% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment1.R
\name{read_datalogger_file}
\alias{read_datalogger_file}
\title{Read in a CSV of accelerometer data produced by datalogger}
\usage{
read_datalogger_file(inputpath)
}
\arguments{
\item{inputpath}{the path to the DataLogger input file.}
}
\value{
A dataframe of accelerometer data.
}
\description{
Read in a CSV of accelerometer data produced by datalogger
}
