% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{runExample}
\alias{runExample}
\title{Run an example algorithm from the class}
\usage{
runExample(example, output_directory)
}
\arguments{
\item{example}{Which example to run}
}
\description{
Run an example algorithm from the class
}
